Backport the 64 bit patch for GET DIAGNOSTICS
Along the way, modify all places where SPI_processed might exceed
2^32 bits.

This backports the following patches:
* 23a27b039d94ba359286694831eafe03cd970eef
* f3f3aae4b7841f4dc51129691a7404a03eb55449
* 74a379b984d4df91acec2436a16c51caee3526af
* ab737f6ba9fc0a26d32a95b115d5cd0e24a63191

+++ contrib/xml2/xpath.c
@@ -522,8 +522,8 @@ xpath_table(PG_FUNCTION_ARGS)
 
 	int			numpaths;
 	int			ret;
-	int			proc;
-	int			i;
+	uint64			proc;
+	uint64			i;
 	int			j;
 	int			rownr;			/* For issuing multiple rows from one original
 								 * document */
@@ -629,7 +629,6 @@ xpath_table(PG_FUNCTION_ARGS)
 			 query_buf.data);
 
 	proc = SPI_processed;
-	/* elog(DEBUG1,"xpath_table: SPI returned %d rows",proc); */
 	tuptable = SPI_tuptable;
 	spi_tupdesc = tuptable->tupdesc;
